Create YAML Header

Creates the R markdown header, and first setup chunk. The setup chunk quiets messages, attaches any needed packages, and loads the object to be printed.

create_yaml_header(object_path, pkg_to_attach, reference_docx)

Arguments

object_path

(path)
path to saved RDS object file.

pkg_to_attach

(character)
character vector of packages that are attached at the top of the R markdown document. These are needed to access the appropraite print methods for the loaded table object.

reference_docx

(path)
path the reference Word document.

Value

YAML character vector

Examples

create_yaml_header(
  object_path = tempfile(fileext = ".rds"),
  pkg_to_attach = "flextable",
  reference_docx = tempfile(fileext = ".docx")
) |>
  cat(sep = "\n")
#> ---
#> output:
#>   word_document:
#>     reference_docx: /tmp/RtmpsCYFR8/file1d012cc581e5.docx
#> ---
#> 
#> ```{r setup, include=FALSE}
#> knitr::opts_chunk$set(echo = FALSE, message = FALSE)
#> library(flextable)
#> x <- readRDS(file = "/tmp/RtmpsCYFR8/file1d012fdd9b1e.rds")
#> if (!inherits(x, 'list')) x <- list(x)
#> ```
#>